Refactor: SRP module split into src/ package with local execution support (#29)
* Initial plan

* Implement strict JSON mode for Gemini response parsing with comprehensive tests


* Address code review: rename private function to public, fix line_num check


* Refactor: move all Python code into src/ package with SRP modules

Split the monolithic entrypoint.py into focused modules:
- src/config.py: Configuration and env var validation
- src/utils.py: General-purpose utilities
- src/review_parser.py: JSON response parsing
- src/prompts.py: Prompt templates
- src/github_client.py: GitHub API interactions
- src/review_formatter.py: Review output formatting
- src/quota.py: Rate limiting and quota tracking
- src/gemini_client.py: Gemini AI service
- src/main.py: CLI entry point and orchestration

Updated Dockerfile, tests, README with local usage docs.


* Address PR review comments: strict JSON mode, separate summarize model, error handling, type fixes, test split

- Add response_mime_type: application/json to enforce strict JSON from Gemini
- Create separate summarize_model without review system_instruction
- Remove duplicate review_prompt from prompt_parts (already in system_instruction)
- Add error checking for GitHub API response in main.py
- Add timeout=30 to requests.post in github_client.py
- Add type annotation for chunk_size parameter in utils.py
- Remove unused List import, use NotRequired for optional TypedDict fields in config.py
- Validate env vars are non-empty (not just non-None) in check_required_env_vars
- Split test_entrypoint.py into test_prompts.py and test_review_formatter.py


---------
