+++ requirements-dev.txt
@@ -12,6 +12,6 @@
 
 -r ./requirements.txt
 
-pytest==4.3.0
+pytest>=7,<9
 
 pre-commit>=3, <4
